fix(ecg): 优化逻辑
fix #2183

+++ Params/ECGParam/ECGParam.h
@@ -408,7 +408,7 @@ private:
     bool _isFristConnect;         // 开机后是否正常连接过导联
     ECGGain _autoGain[ECG_LEAD_NR];      // 自动增益计算出的实际增益
 
-    unsigned getMaxHrValue();
-    unsigned getMinHrValue();
+    short getMaxHrValue();
+    short getMinHrValue();
 };
 #define ecgParam (ECGParam::getInstance())
